% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R
\name{model1}
\alias{model1}
\title{Simple regression model}
\usage{
model1(x, y)
}
\arguments{
\item{x}{an independent variable}

\item{y}{a dependent or response variable}
}
\value{
regression coeficient
}
\description{
Simple regression model
}
\examples{
x <- c(1,3,4,5,8,10)
y <- c(4,6,8,7,10,16)
model1(x,y)
}
