% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{regmodel}
\alias{regmodel}
\title{Classical linear regression  model}
\usage{
regmodel(X, y)
}
\arguments{
\item{X}{a matrix of independent variables with dimension \code{nxp}, where \code{p} is the number of independent variable (\code{k}) + 1}

\item{y}{a vector of dependent variable (or response variable) with dimension \code{nx1}}
}
\value{
\code{B} is a matrix of regression coefficients with dimension \code{px1}, where
the first element is intercept (\code{b0})
the second, the third,... be regression coefficient corresponding to \code{x1, x2,...,xk}, respectively.
}
\description{
To fit simple and multiple linear regression  model
}
\examples{
x1 <- c(25,40,50,45,35,28,34,52,38,41,27,50)
x2 <- c(4,1,0.5,0.5,2,4.5,2,0.5,2.6,1,4.5,0.5)
y <- c(3,4.2,4.8,4.5,4.5,3.5,4,5.5,4.6,5,4,4.8)
X <- as.matrix(data.frame(x0=1,x1,x2))
regmodel(X,y)
}
\references{
Myers, Raymond H.; Milton, Janet S.(1991) A First Course in the Theory of Linear Statistical Models. PWS-KENT Pub. Co, Boston.
}
